requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
python-binance>=1.0.19
ccxt>=4.1.0
web3>=6.11.0
aiohttp>=3.9.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
cryptography>=41.0.0
flask>=3.0.0
flask-cors>=4.0.0
colorama>=0.4.6
click>=8.1.0
schedule>=1.2.0
sqlalchemy>=2.0.0
redis>=5.0.0
celery>=5.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

